% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGraph.R
\name{get_graph_info}
\alias{get_graph_info}
\title{Get travel path of graph}
\usage{
get_graph_info(edgeMatrix = HGraph::edge, varVec = c("a", "b", "c"))
}
\arguments{
\item{edgeMatrix}{a matrix}

\item{varVec}{a vector}
}
\value{
Graph struct
}
\description{
Get travel path of graph
}
\examples{
aedge <- matrix(0, 3, 3, dimnames=list(c("a", "b", "c"), c("a", "b", "c")))
aedge["a","b"]<-1
aedge["a","c"]<-1
aedge["b","c"]<-1
results <- get_graph_info(edgeMatrix=aedge, varVec=c("a", "b", "c"))
print(results)
}
