% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGraph.R
\docType{data}
\name{vec}
\alias{vec}
\title{Get travel path of graph}
\format{
An object of class \code{character} of length 3.
}
\usage{
vec
}
\arguments{
\item{edgeMatrix}{varVec}
}
\value{
Graph struct
}
\description{
Get travel path of graph
}
\examples{
get_graph(edgeMatrix,varVec)
}
\keyword{datasets}
